requirements.txt
Flask==2.3.2
openai==0.27.8
google-generativeai==0.3.2
resend==0.5.2
pymongo==3.12.0
pymongo[srv]==3.12.0
flask-cors==3.0.10
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
